Migrations enhancements

+++ src/database/migrations/20240212142420_categories.js
@@ -38,7 +38,7 @@ module.exports = {
 					allowNull: false
 				},
 				limit: {
-					type: sequelize.DataTypes.INTEGER,
+					type: sequelize.DataTypes.DOUBLE,
 					allowNull: false
 				},
 				createdAt: {
